[POSUI-244] BroadPOS payment app's security keyboard is covering the POSUI apps toast error message for incorrect password entry

+++ app/src/main/java/com/paxus/pay/poslinkui/demo/entry/text/fleet/FleetFragment.java
@@ -138,7 +138,7 @@ public class FleetFragment extends BaseEntryFragment {
     }
 
     @Override
-    protected void onEntryDeclined(long errCode, String errMessage) {
+    protected void onEntryDeclined(int errCode, String errMessage) {
         super.onEntryDeclined(errCode, errMessage);
 
         //re enter if next request declined
